Split into multiple files in roomba
